refactor: improve operation order components
